able to change password and so on

+++ web/ts/ShowDetail.ts
@@ -26,18 +26,19 @@ class ShowDetail {
             $("#" + ShowDetail.NAME + "-commentsbtn").click(ShowDetail.postComments);
             $("#" + ShowDetail.NAME + "-scommentsbtn").click(ShowDetail.showComments);
             $("#" + ShowDetail.NAME + "-likebtn").click(ShowDetail.likePost);
+            $("#" + ShowDetail.NAME + "-dislikebtn").click(ShowDetail.dislikePost);
             $("#" + ShowDetail.NAME + "-editbtn").click(ShowDetail.clickEdit);
             $("#" + ShowDetail.NAME + "-delbtn").click(ShowDetail.clickDelete);
             $("#" + ShowDetail.NAME + "-closebtn").click(ShowDetail.hide);
         }
     }
 
-    private static reloadLike() {
+    private static reload() {
         let id = "" + $("#" + ShowDetail.NAME + "-detailId").val();
         // Issue a GET, and then pass the result to update()
         $.ajax({
-            type: "GET",
-            url: "/messages/" + id + "/likes",
+            type: "POST",
+            url: "/messages/" + id,
             dataType: "json",
             data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
             success: ShowDetail.likeResponse
@@ -47,7 +48,10 @@ class ShowDetail {
     private static likeResponse(data: any) {
         $("#" + ShowDetail.NAME + "-likebtn").text("");
         $("#" + ShowDetail.NAME + "-likebtn").text("Like: " + data.mData.mLikes);
+        $("#" + ShowDetail.NAME + "-dislikebtn").text("");
+        $("#" + ShowDetail.NAME + "-dislikebtn").text("Like: " + data.mData.mDislikes);
     }
+
 
     /**
      * Refresh() doesn't really have much meaning, but just like in sNavbar, we
@@ -70,6 +74,7 @@ class ShowDetail {
         $("#" + ShowDetail.NAME + "-commentsbtn").click(ShowDetail.postComments);
         $("#" + ShowDetail.NAME + "-scommentsbtn").click(ShowDetail.showComments);
         $("#" + ShowDetail.NAME + "-likebtn").click(ShowDetail.likePost);
+        $("#" + ShowDetail.NAME + "-dislikebtn").click(ShowDetail.dislikePost);
         $("#" + ShowDetail.NAME + "-editbtn").click(ShowDetail.clickEdit);
         $("#" + ShowDetail.NAME + "-delbtn").click(ShowDetail.clickDelete);
         $("#" + ShowDetail.NAME + "-closebtn").click(ShowDetail.hide);
@@ -87,6 +92,7 @@ class ShowDetail {
         $("#" + ShowDetail.NAME + "-detailId").val("");
         $("#" + ShowDetail.NAME + "-detailPostUid").val("");
         $("#" + ShowDetail.NAME + "-created").text("");
+        $("#" + ShowDetail.NAME + "-dislikebtn").text("");
         $("#" + ShowDetail.NAME + "-likebtn").text("");
         $("#" + ShowDetail.NAME).modal("hide");
     }
@@ -99,18 +105,6 @@ class ShowDetail {
      */
     public static show(data: any) {
         $("#" + ShowDetail.NAME + "-detailId").val(data.mData.mId);
-        /*
-        let id = "" + $("#" + ShowDetail.NAME + "-detailId").val();
-        $.ajax({
-            type: "POST",
-            url: "/" + id + "/listcomments",
-            dataType: "json",
-            data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
-            // 用里的show
-            success: ShowDetail.update
-            //如果refresh 不行就改回listAllComments
-        });
-        */
         $("#" + ShowDetail.NAME + "-title").text(data.mData.mTitle);
         $("#" + ShowDetail.NAME + "-username").text(data.mData.mName);
         $("#" + ShowDetail.NAME + "-message").val(data.mData.mContent);
@@ -118,6 +112,7 @@ class ShowDetail {
         $("#" + ShowDetail.NAME + "-detailPostUid").val(data.mData.uId);
         $("#" + ShowDetail.NAME + "-created").text(data.mData.mCreated);
         $("#" + ShowDetail.NAME + "-likebtn").text("Like: " + data.mData.mLikes);
+        $("#" + ShowDetail.NAME + "-dislikebtn").text("Dislike: " + data.mData.mDislikes);
         $("#" + ShowDetail.NAME).modal("show");
     }
 
@@ -137,8 +132,23 @@ class ShowDetail {
             type: "PUT",
             url: "/messages/" + id + "/likes",
             dataType: "json",
-            data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
-            success: ShowDetail.reloadLike
+            data: JSON.stringify({ uEmail: uemail, sessionKey: ukey , uid: uid}),
+            success: ShowDetail.reload
+        });
+    }
+
+    private static dislikePost() {
+        // get the values of the id of the current post
+        let id = "" + $("#" + ShowDetail.NAME + "-detailId").val();
+
+        // set up an AJAX post.  When the server replies, the result will go to
+        // onSubmitResponse
+        $.ajax({
+            type: "PUT",
+            url: "/messages/" + id + "/dislikes",
+            dataType: "json",
+            data: JSON.stringify({ uEmail: uemail, sessionKey: ukey , uid: uid}),
+            success: ShowDetail.reload
         });
     }
 
@@ -167,6 +177,10 @@ class ShowDetail {
     private static postComments(){
         let mid = "" + $("#" + ShowDetail.NAME + "-detailId").val();
         let text = "" + $("#" + ShowDetail.NAME + "-postcomments").val();
+        if (text === "") {
+            window.alert("Error: comment is not valid");
+            return;
+        }
         $.ajax({
             type: "POST",
             url: "/" + mid + "/comments",
@@ -184,34 +198,36 @@ class ShowDetail {
         // for now, just print the ID that goes along with the data in the row
         // whose "delete" button was clicked
         let id = "" + $("#" + ShowDetail.NAME + "-detailId").val();
-
-        ShowDetail.hide();
-        $.ajax({
-            type: "DELETE",
-            url: "/messages/" + id,
-            dataType: "json",
-            data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
-            // TODO: we should really have a function that looks at the return
-            //       value and possibly prints an error message.
-            success: ElementList.refresh
-        });
+        if (Number(id) == uid) {
+            ShowDetail.hide();
+            $.ajax({
+                type: "DELETE",
+                url: "/messages/" + id,
+                dataType: "json",
+                data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
+                // TODO: we should really have a function that looks at the return
+                //       value and possibly prints an error message.
+                success: ElementList.refresh
+            });
+        }
     }
 
     /**
-     * clickEdit is the code we run in response to a click of a delete button
+     * clickEdit is the code we run in response to a click of a edit button
      */
     private static clickEdit() {
         // as in clickDelete, we need the ID of the row
         let id = "" + $("#" + ShowDetail.NAME + "-detailId").val();
-
-        ShowDetail.hide();
-        $.ajax({
-            type: "GET",
-            url: "/messages/" + id,
-            dataType: "json",
-            data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
-            success: EditEntryForm.show
-        });
+        if (Number(id) == uid) {
+            ShowDetail.hide();
+            $.ajax({
+                type: "POST",
+                url: "/messages/" + id,
+                dataType: "json",
+                data: JSON.stringify({ uEmail: uemail, sessionKey: ukey }),
+                success: EditEntryForm.show
+            });
+        }
     }
 
     private static other() {
